Move test suite to its own module, clean up Main
We now have a more sensible way of determining what to do on startup.
Before, we ran the tests, and if there were any failures, report them.
If not, parse and print the AST for stdin. This was pretty gross; now
you can run the tests by passing "test" as the first argument.