Emit assembly for a whole program
...including the entry point declaration, and an exit syscall at the
end. We're just exiting with whatever's in r0 right now, which is the
result of the last thing evaluated. This will be helpful as an early
mechanism for testing, but obviously we want to remove it at some point.

Unsurprisingly, this is segfaulting. My guess is the variable references
via fp don't work the way I guessed (really, guessed, didn't even try to
look up), so that will need fixing.